Fix round number for wednesday match in 2019
As with other recent bugs for footywire round number calculations,
I'm a bit confused as to what changed regarding how weeks and
weekdays are counted to cause this, but something moved Wednesdays
from the beginning to the end of the round week by default.

Changing the between filter to use Sunday and Tuesday as the limits
makes the default round week Wednesday to Tuesday as intended.