Add Match.ID column to guarantee unique ID columns for pivot
With the limited number of venues for the current rounds, the
AFL has scheduled two matches for the same venue on the same day,
which breaks the uniqueness assumption for those value combinations.
Adding an ID column to facilitate pivoting fixes the error raised
from duplication causing team columns to have vectors of names
instead of individual team names.