Calculate round numbers based on scraped round names for betting odds
This finishes replacement of calculation of rounds based on
calendar weeks.

I moved some of the data processing to the per-season loops
earlier in the function to make it easier to calculate the max
regular season round per season.

I added a fix to make sure the valid start season isn't greater
than the valid end season, which we weren't checking before.