Handle inconsistent column data types
AFLTables is returning some columns (e.g. Round, Jumper.No.)
as character type for the 2020 season, but as double type for 2021,
which raises an error when we try to join the season
dataframes. We have to filter out empty dataframes before setting
the types, because they don't have columns to convert.

I added '2021' to the tests for this function to avoid regressions.