Return NULL if season_id doesn't exist
The AFLM season IDs only go back to 2012, so we don't expect to
be able to fetch season info for the 2000 season. The tests
for fetch_ladder_afl reflect this, but we were only checking for
a blank round_id for the early return, which caused an error
later on for the missing season data.