fixes #168
There was an issue with some of the cached ID's. This updates the ids for new seasons. Doesn't fix older issues that I'll have to look at but fixes this specific example.

Also requires an update of cached data which I've pushed at the same time (may not quite be updated)